Transfer fix and More swappable controllers
The DB code is updated so that controllers that have a connection to the right table can connect to a node and pull data successfully from it. Before, the issue was the headers were rewritten into the table too but now the new Controller will get the headers of the table before it starts running so that mistake won't be made for future node connections.